backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
python-multipart>=0.0.9

asyncpg>=0.29.0
pytz>=2024.1
pydantic-settings>=2.0.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
requests>=2.31.0
google-genai>=1.0.0
httpx>=0.27.0
psycopg2-binary>=2.9.9
google-cloud-bigquery>=3.0.0
